Update de.json
Added some corrections. I not sure about the capitals for weak. Is this the one in the RPE? There are only capitals letters. In this corrected version were both kinds of letters. SCHLECHT is very poor/weak. As well it is ok to say SEHR SCHWACH as a harder wording for poor SCHWACH (very poor). Schlecht is a little more to feeling bad, so i think one word for the best understanding. Great work....