Update nl.json (some random readability changes)
I rewrote a lot of sentences that were written in a correct, but less readable way. (active vs passive etc.) This makes the sentences more readable and understandable. 

I also rewrote all sentences in the less formal form (jij/je) instead of using the formal form (U/uw). This wasn't consequently done. 

If you want all sentences back in the formal form, I'll change that, but is not in line how intervals present itself to their users, because it creates a certain "distance".